Revert "Revert "VPAAMP-201 Move SpeedCache to common AampSpeedCache.h header …"
This reverts commit d09a62c881d4fb6102adc59be57f0b6eab354fe0.
