+++ support/aampabr/HybridABRManager.h
@@ -27,6 +27,7 @@
 #include <string>
 #include <cstdio>
 #include "ABRManager.h"
+#include "AampSpeedCache.h"
 
 class HybridABRManager:public ABRManager
 {
